refactor: Delete and rearrange files as necessary
