+++ src/test/java/com/ibm/cloud/sdk/core/test/BaseServiceUnitTest.java
@@ -1,35 +1,18 @@
-/**
- * Copyright 2017 IBM Corp. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package com.ibm.cloud.sdk.core.test;
-
-import java.io.IOException;
-
-import org.apache.commons.lang3.StringUtils;
-import org.junit.After;
+package com.ibm.cloud.sdk.core.test.service;
 
 import com.google.gson.Gson;
 import com.ibm.cloud.sdk.core.http.HttpMediaType;
 import com.ibm.cloud.sdk.core.util.GsonSingleton;
-
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
+import org.apache.commons.lang3.StringUtils;
+import org.junit.After;
 
-/**
- * Utility class to Mock the Watson Services.
- *
- */
-public abstract class WatsonServiceUnitTest extends WatsonServiceTest {
+import java.io.IOException;
 
+import static com.ibm.cloud.sdk.core.http.HttpHeaders.CONTENT_TYPE;
+
+public class BaseServiceUnitTest {
   /** The Constant DELETE. */
   protected static final String DELETE = "DELETE";
 
@@ -86,5 +69,4 @@ public abstract class WatsonServiceUnitTest extends WatsonServiceTest {
   protected static MockResponse jsonResponse(Object body) {
     return new MockResponse().addHeader(CONTENT_TYPE, HttpMediaType.APPLICATION_JSON).setBody(GSON.toJson(body));
   }
-
 }
