--- src/test/resources/config.properties
@@ -1,62 +1,0 @@
-# SERVICE CREDENTIALS
-
-alchemy.alchemy=API_KEY
-
-conversation.username=SERVICE_USERNAME
-conversation.password=SERVICE_PASSWORD
-conversation.url=SERVICE_URL
-conversation.workspace_id=PLACEHOLDER
-
-dialog.dialog_id=PLACEHOLDER
-dialog.password=SERVICE_PASSWORD
-dialog.url=SERVICE_URL
-dialog.username=SERVICE_USERNAME
-
-discovery.password=SERVICE_PASSWORD
-discovery.url=SERVICE_URL
-discovery.username=SERVICE_USERNAME
-
-document_conversion.password=SERVICE_PASSWORD
-document_conversion.url=SERVICE_URL
-document_conversion.username=SERVICE_USERNAME
-
-language_translator.model_id=PLACEHOLDER
-language_translator.password=SERVICE_PASSWORD
-language_translator.url=SERVICE_URL
-language_translator.username=SERVICE_USERNAME
-
-natural_language_classifier.classifier_id=PLACEHOLDER
-natural_language_classifier.password=SERVICE_PASSWORD
-natural_language_classifier.url=SERVICE_URL
-natural_language_classifier.username=SERVICE_USERNAME
-
-personality_insights.password=SERVICE_PASSWORD
-personality_insights.url=SERVICE_URL
-personality_insights.username=SERVICE_USERNAME
-
-retrieve_and_rank.cluster_id=PLACEHOLDER
-retrieve_and_rank.config_name=PLACEHOLDER
-retrieve_and_rank.password=SERVICE_PASSWORD
-retrieve_and_rank.ranker_id=PLACEHOLDER
-retrieve_and_rank.url=SERVICE_URL
-retrieve_and_rank.username=SERVICE_USERNAME
-
-speech_to_text.password=SERVICE_PASSWORD
-speech_to_text.url=SERVICE_URL
-speech_to_text.username=SERVICE_USERNAME
-
-text_to_speech.voice_name=PLACEHOLDER
-text_to_speech.password=SERVICE_PASSWORD
-text_to_speech.url=SERVICE_URL
-text_to_speech.username=SERVICE_USERNAME
-
-tone_analyzer.v3.password=SERVICE_PASSWORD
-tone_analyzer.v3.url=SERVICE_URL
-tone_analyzer.v3.username=SERVICE_USERNAME
-
-tradeoff_analytics.password=SERVICE_PASSWORD
-tradeoff_analytics.url=SERVICE_URL
-tradeoff_analytics.username=SERVICE_USERNAME
-
-visual_recognition.v3.api_key=API_KEY
-visual_recognition.v3.url=SERVICE_URL
